refactor(character): split CombatRow and WalkSpeedRow into separate files
